chore(sanity): move `ConfirmDeleteDialog` to legacy path
This makes way for the new deletion experience; coexisting with the
legacy experience allows toggling between the experiences if necessary.
