+++ packages/sanity/src/structure/components/confirmDeleteDialog/legacy/ConfirmDeleteDialogBody.styles.tsx
@@ -3,8 +3,8 @@ import {Box, Flex, Inline, rem, Text} from '@sanity/ui'
 import {useTranslation} from 'sanity'
 import {styled} from 'styled-components'
 
-import {Tooltip} from '../../../ui-components'
-import {structureLocaleNamespace} from '../../i18n'
+import {Tooltip} from '../../../../ui-components'
+import {structureLocaleNamespace} from '../../../i18n'
 
 export const ChevronWrapper = styled(Box)`
   margin-left: auto;
