+++ packages/sanity/src/structure/components/confirmDeleteDialog/legacy/ConfirmDeleteDialogLegacy.tsx
@@ -3,9 +3,9 @@ import {useCallback, useId, useMemo} from 'react'
 import {getPublishedId, LoadingBlock, useDocumentVersions, useTranslation} from 'sanity'
 import {styled} from 'styled-components'
 
-import {Dialog} from '../../../ui-components'
-import {structureLocaleNamespace} from '../../i18n'
-import {DocTitle} from '../DocTitle'
+import {Dialog} from '../../../../ui-components'
+import {structureLocaleNamespace} from '../../../i18n'
+import {DocTitle} from '../../DocTitle'
 import {ConfirmDeleteDialogBody} from './ConfirmDeleteDialogBody'
 import {useReferringDocuments} from './useReferringDocuments'
 
@@ -49,7 +49,7 @@ export interface ConfirmDeleteDialogProps {
  *
  * @internal
  */
-export function ConfirmDeleteDialog({
+export function ConfirmDeleteDialogLegacy({
   id,
   type,
   action = 'delete',
