+++ packages/sanity/src/structure/components/confirmDeleteDialog/legacy/VersionsPreviewList.tsx
