+++ packages/sanity/src/structure/components/confirmDeleteDialog/legacy/__tests__/index.test.tsx
